# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Velora DNS Management API",
    "version": "1.0.0",
    "description": "Session cookies require X-CSRF-Token on mutations. API bearer tokens bypass CSRF and require the documented scope. All API responses use a data or error envelope except text/dns zone export. Secrets are never returned by read endpoints."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "paths": {
...
    "/api/v1/clients": {
      "get": {
        "operationId": "get_clients",
        "summary": "Get clients",
        "tags": [
          "clients"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ClientList"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "x-required-role": "viewer-or-higher"
      },
      "post": {
        "operationId": "post_clients",
        "summary": "Post clients",
        "tags": [
          "clients"
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Client"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "x-required-role": "operator-or-admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientInput"
              }
            }
          }
        }
      }
    },
    "/api/v1/clients/{id}": {
      "delete": {
        "operationId": "delete_clients_id",
        "summary": "Delete clients/{id}",
        "tags": [
          "clients"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Client"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "x-required-role": "operator-or-admin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ]
      },
      "put": {
        "operationId": "put_clients_id",
        "summary": "Put clients/{id}",
        "tags": [
          "clients"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Client"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "x-required-role": "operator-or-admin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientInput"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/policies": {
      "get": {
        "operationId": "get_policies",
        "summary": "Get policies",
        "tags": [
          "policies"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Policy"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "x-required-role": "viewer-or-higher"
      },
      "post": {
        "operationId": "post_policies",
        "summary": "Post policies",
        "tags": [
          "policies"
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Policy"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "x-required-role": "operator-or-admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyInput"
              }
            }
          }
        }
      }
    },
    "/api/v1/policies/effective": {
      "get": {
        "operationId": "get_policies_effective",
        "summary": "Get policies/effective",
        "tags": [
          "policies"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EffectivePolicy"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "x-required-role": "viewer-or-higher"
      }
    },
    "/api/v1/policies/{id}": {
      "delete": {
        "operationId": "delete_policies_id",
        "summary": "Delete policies/{id}",
        "tags": [
          "policies"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Policy"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "x-required-role": "operator-or-admin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ]
      },
      "put": {
        "operationId": "put_policies_id",
        "summary": "Put policies/{id}",
        "tags": [
          "policies"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Policy"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "x-required-role": "operator-or-admin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyInput"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "velora_ API token; scopes read, write, or admin"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "velora_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Required for cookie-authenticated mutations"
      }
    },
    "schemas": {
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Configuration path such as dns.allowed_clients[1]"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "description": "Invalid or restart-only configuration fields (invalid_config, config_requires_restart)"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Client": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "addresses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "group": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "addresses",
          "enabled"
        ]
      },
      "ClientInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "addresses": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "maxItems": 16,
            "description": "IP addresses or CIDR networks"
          },
          "group": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "addresses"
        ]
      },
      "Policy": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "client_id": {
            "type": "integer",
            "format": "int64"
          },
          "mode": {
            "type": "string",
            "enum": [
              "default",
              "disabled",
              "custom"
            ]
          },
          "blocklists": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Blocklist source IDs used in custom mode; they apply even when disabled globally"
          },
          "allow": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "block": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "client_id",
          "mode",
          "blocklists",
          "allow",
          "block",
          "enabled"
        ]
      },
      "PolicyInput": {
        "type": "object",
        "properties": {
          "client_id": {
            "type": "integer",
            "format": "int64"
          },
          "mode": {
            "type": "string",
            "enum": [
              "default",
              "disabled",
              "custom"
            ]
          },
          "blocklists": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "allow": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Domains (and their subdomains) never blocked for this client"
          },
          "block": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Extra domains (and their subdomains) blocked for this client"
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "client_id",
          "mode"
        ]
      },
      "EffectivePolicy": {
        "type": "object",
        "properties": {
          "client": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Client"
              },
              {
                "type": "null"
              }
            ]
          },
          "policy": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Policy"
              },
              {
                "type": "null"
              }
            ]
          },
          "mode": {
            "type": "string",
            "enum": [
              "default",
              "disabled",
              "custom"
            ]
          }
        },
        "required": [
          "client",
          "policy",
          "mode"
        ]
      },
...
      "ClientActivity": {
        "type": "object",
        "properties": {
          "client_ip": {
            "type": "string"
          },
          "queries": {
            "type": "integer",
            "format": "int64"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "client_ip",
          "queries",
          "last_seen"
        ]
      },
      "ClientList": {
        "type": "object",
        "properties": {
          "clients": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "format": "int64"
                },
                "name": {
                  "type": "string"
                },
                "addresses": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "group": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "enabled": {
                  "type": "boolean"
                },
                "activity": {
                  "type": [
                    "object",
                    "null"
                  ],
                  "properties": {
                    "queries": {
                      "type": "integer",
                      "format": "int64"
                    },
                    "last_seen": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    }
                  },
                  "description": "Last 24 hours of retained queries; null when query logging is disabled"
                }
              }
            }
          },
          "unnamed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientActivity"
            }
          }
        },
        "required": [
          "clients",
          "unnamed"
        ]
      },
...
    }
  }
}